+++ src/js/common/schemaform/reducers/index.js
@@ -2,14 +2,14 @@ import _ from 'lodash/fp';
 import { getDefaultFormState } from 'react-jsonschema-form/lib/utils';
 
 import {
-  setHiddenFields,
-  removeHiddenData,
-  updateRequiredFields,
   createFormPageList,
-  updateSchemaFromUiSchema,
+} from '../helpers';
+
+import {
+  updateSchemaData,
   replaceRefSchemas,
   updateItemsSchema
-} from '../helpers';
+} from '../formState';
 
 import { SET_DATA,
   SET_EDIT_MODE,
@@ -35,17 +35,8 @@ function recalculateSchemaAndData(initialState) {
 
       // Recalculate any required fields, based on the new data
       const page = state[pageKey];
-      let schema = updateItemsSchema(page.schema, formData);
-      schema = updateRequiredFields(schema, page.uiSchema, formData);
 
-      // Update the schema with any fields that are now hidden because of the data change
-      schema = setHiddenFields(schema, page.uiSchema, formData);
-
-      // Update the schema with any general updates based on the new data
-      schema = updateSchemaFromUiSchema(schema, page.uiSchema, page.data, state);
-
-      // Remove any data that's now hidden in the schema
-      const data = removeHiddenData(schema, page.data);
+      const { data, schema } = updateSchemaData(page.schema, page.uiSchema, formData, page.data, state);
 
       if (page.data !== data || page.schema !== schema) {
         const newPage = _.assign(page, {
